+++ TheEngine/src/Components/TilemapComponent.cpp
@@ -74,7 +74,7 @@ std::string Core::TilemapComponent::GetNameAttributeFromLine(const std::string& 
 	size_t NameEndPos = Line.find("\"", NameStartPos);
 
 	std::string Name;
-	for (int i = NameStartPos; i < NameEndPos; i++)
+	for (size_t i = NameStartPos; i < NameEndPos; i++)
 	{
 		Name += Line[i];
 	}
@@ -152,7 +152,7 @@ std::string Core::TilemapComponent::GetStringValueFromFlag(const std::string& Li
 std::string Core::TilemapComponent::GetDelimitedStringValue(const std::string& Line, const size_t Begin, const size_t End) const
 {
 	std::string WidthStr;
-	for (int i = Begin; i < End; i++)
+	for (size_t i = Begin; i < End; i++)
 	{
 		WidthStr += Line[i];
 	}
@@ -243,7 +243,7 @@ void Core::TilemapComponent::Draw()
 				if (TileId >= T._Sources.size()) continue;
 
 				Rect<int> Src = T._Sources[TileId - 1];
-				Rect<int> Dst(x * mScaledTileSize.X, y * mScaledTileSize.Y, mScaledTileSize.X, mScaledTileSize.Y);
+				Rect<int> Dst(x * (int)mScaledTileSize.X, y * (int)mScaledTileSize.Y, (int)mScaledTileSize.X, (int)mScaledTileSize.Y);
 
 				Graphic().DrawTexture(T._ImageId, Src, Dst, 0.0, Flip::None, Color::White);
 			}
@@ -287,8 +287,8 @@ void Core::TilemapComponent::SetTilemapData(const std::string& TiledFile, float 
 	Graphic().GetWindowSize(&WindowW, &WindowH);
 	mInitialTileSize.X = SrcTileW;
 	mInitialTileSize.Y = SrcTileH;
-	mScaledTileSize.X = WindowW / mTileCount.X;
-	mScaledTileSize.Y = WindowH / mTileCount.Y;
+	mScaledTileSize.X = (float)(WindowW / mTileCount.X);
+	mScaledTileSize.Y = (float)(WindowH / mTileCount.Y);
 }
 
 void Core::TilemapComponent::Destroy()
